Commit 1 CPL
Añadido conteo de "Me gusta" y "no me gusta" en preguntas.

Se actualiza también para respuestas y comentarios, pero no se visualiza al refrescar la página.

+++ src/main/java/com/cplcursos/java/kosso/repositories/PuntosForoRepo.java
@@ -9,8 +9,8 @@ public interface PuntosForoRepo extends JpaRepository<PuntosForo, Long> {
     Long countByIdContenidoAndTipoContenido(Long idContenido, String tipoContenido);
 
     @Query("SELECT COUNT(a) from PuntosForo a WHERE a.puntos > 0 AND a.idContenido = ?1 AND a.tipoContenido = ?2")
-    Integer cuentaLike(Long idContenido, String tipoContenido);
+    Long cuentaLike(Long idContenido, String tipoContenido);
 
     @Query("SELECT COUNT(a) from PuntosForo a WHERE a.puntos < 0 AND a.idContenido = ?1 AND a.tipoContenido = ?2")
-    Integer cuentaDislike(Long idContenido, String tipoContenido);
+    Long cuentaDislike(Long idContenido, String tipoContenido);
 }
